+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.1]
+### Changed
+- Changed Control class property renderNoContent to be an init parameter so that it can be modified as a tag attribute, e.g. <cms:Content renderNoContent="true|false" />
+### Fixed
+- Minor path issues when not using DIRECTORY_SEPARATOR constant.
 ## [1.2.0]
 ### Added
 - New WebImage\Config\Config for better configuration merging capabilities.